make VersionCheckUtils class
