+++ tools/local-dev.sh
@@ -22,11 +22,11 @@ terra config set image --default
 
 echo "Pulling the default Docker image"
 defaultDockerImage=$(terra config get image)
-#docker pull $defaultDockerImage
+docker pull $defaultDockerImage
 
 echo "Setting the server to its current value, to pull any changes"
 currentServer=$(terra config get server)
-#terra config set server --name=$currentServer
+terra config set server --name=$currentServer
 
 echo "Making all 'tools' scripts executable"
 chmod a+x tools/*
